packages/example-3-2-serverless-azure-function/pyproject.toml
[project]
name = "example-3-2-serverless-azure-function"
version = "1.0.0"
description = "Step 3.2 — alt-python/boot Azure Functions serverless adapter"
requires-python = ">=3.12"
dependencies = [
    "alt-python-boot-lib",
    "alt-python-boot-azure-function",
    "azure-functions>=1.18",
]

[tool.uv.sources]
alt-python-boot-lib = { workspace = true }
alt-python-boot-azure-function = { workspace = true }
